services/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
redis==5.0.1
nats-py==2.7.2
PyJWT==2.8.0
